+++ database/migrations/2024_01_30_153023_change_pamyra_orders.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('tms_pamyra_orders', function (Blueprint $table) {
-            $table->dropColumn('payment_method');
+            if (Schema::hasColumn('tms_pamyra_orders', 'payment_method')) {
+                $table->dropColumn('payment_method');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('tms_pamyra_orders', function (Blueprint $table) {
-            $table->string('payment_method')->nullable()->change();
+            if (Schema::hasColumn('tms_pamyra_orders', 'payment_method')) {
+                $table->string('payment_method')->nullable()->change();
+            }
         });
     }
 };
